pyproject.toml
[tool.poetry]
name = "ipblocklistservice"
readme = "README.md"
authors = []
description = "IP Blocklist Service"
version = "1"

[tool.poetry.dependencies]
python = "^3.10.4"
flask = "^3.0"
gunicorn = "^21.2"
redis = "^5.0.1"
flask-swagger-ui = "^4.11.1"
pytest = "^8.0.2"
schedule = "^1.2.1"
pytest-env = "^1.1.3"
requests-mock = "^1.11.0"
responses = "^0.25.0"
injector = "^0.21.0"
apscheduler = "^3.10.4"
gitpython = "^3.1.42"
pygithub = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
flake8-pyproject = "^1.2.2"
toml = "^0.10.2"

[tool.bandit]
exclude_dirs = [".venv","tests"]
skips = ["B104"]

[tool.flake8]
ignore = ["F401","E402"]
exclude = [".venv"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest_env]
REDIS_DISABLED = true
